#164 Refactoring folder structure
