+++ src/app.html
@@ -6,7 +6,7 @@
       <nav-bar router.bind="router"></nav-bar>
       <div class="jumbotron hero-jumbotron">
         <div class="container-fluid">
-          <compose view-model="modules/project/search/search-bar" model.bind="router.currentInstruction.fragment == '/'"></compose>
+          <compose view-model="search/search-bar" model.bind="router.currentInstruction.fragment == '/'"></compose>
         </div>
       </div>
     </div>
